+++ Sources/ApptentiveKit/Interactions/Dialog/UI/DialogView.swift
@@ -44,46 +44,85 @@ public class DialogView: UIView {
     /// The color of the separators.
     @objc public dynamic var separatorColor: UIColor = .clear
 
+    /// The amount by which to inset the header image from view's edges and the title label.
+    @objc public dynamic var imageInset: UIEdgeInsets = .zero {
+        didSet {
+            self.headerImageViewTopConstraint.constant = imageInset.top
+            self.headerImageViewLeadingConstraint.constant = self.traitCollection.layoutDirection == .leftToRight ? imageInset.left : imageInset.right
+            self.headerImageViewTrailingConstraint.constant = self.traitCollection.layoutDirection == .leftToRight ? imageInset.right : imageInset.left
+            self.headerImageViewBottomConstraint.constant = imageInset.bottom
+
+            if imageInset.bottom != 0 {
+                NSLayoutConstraint.activate([self.headerImageViewBottomConstraint])
+            } else {
+                NSLayoutConstraint.deactivate([self.headerImageViewBottomConstraint])
+            }
+        }
+    }
+
     let blurEffect: UIBlurEffect
     let backgroundView: UIVisualEffectView
     let headerImageView: UIImageView
+    let headerImageAlternateLabel: UILabel
     let textScrollView: UIScrollView
     let textContentView: UIView
     let titleLabel: UILabel
     let messageLabel: UILabel
     let gradientView: GradientView
-    let middleSeparator: UIVisualEffectView
+    let mainSeparator: UIVisualEffectView
     var buttonSeparators = [UIVisualEffectView]()
     let buttonStackView: UIStackView
 
     var headerImageViewHeightConstraint = NSLayoutConstraint()
+    var headerImageViewAspectConstraint = NSLayoutConstraint()
+    var headerImageViewTopConstraint = NSLayoutConstraint()
+    var headerImageViewLeadingConstraint = NSLayoutConstraint()
+    var headerImageViewTrailingConstraint = NSLayoutConstraint()
+    var headerImageViewBottomConstraint = NSLayoutConstraint()
     var buttonStackViewTopConstraint = NSLayoutConstraint()
     var buttonStackViewLeftConstraint = NSLayoutConstraint()
     var buttonStackViewRightConstraint = NSLayoutConstraint()
     var buttonStackViewBottomConstraint = NSLayoutConstraint()
     var titleBottomConstraint = NSLayoutConstraint()
     var messageBottomConstraint = NSLayoutConstraint()
+    var topSpacingIdealConstraint = NSLayoutConstraint()
+    var topSpacingMinConstraint = NSLayoutConstraint()
+    var topSpacingMaxConstraint = NSLayoutConstraint()
 
-    var isMessageLabelHidden = false {
+    var topSpacingIdealConstraintConstant: CGFloat = 36
+    var topSpacingMinConstraintConstant: CGFloat = 22.3333
+    var topSpacingMaxConstraintConstant: CGFloat = 36
+    var titleBottomConstraintConstant: CGFloat = 24
+    var messageBottomConstraintConstant: CGFloat = 24
+
+    var collapseImageBottomPadding: Bool = false
+    var dialogType: DialogViewModel.DialogType
+
+    var isMessageHidden = false {
         didSet {
-            self.messageLabel.isHidden = self.isMessageLabelHidden
-            self.titleBottomConstraint.isActive = self.isMessageLabelHidden
-            self.messageBottomConstraint.isActive = !self.isMessageLabelHidden
+            self.messageLabel.isHidden = self.isMessageHidden
+            self.titleBottomConstraint.isActive = self.isMessageHidden
+            self.messageBottomConstraint.isActive = !self.isMessageHidden
         }
     }
 
-    override init(frame: CGRect) {
-
+    init(frame: CGRect, collapseImageBottomPadding: Bool, dialogType: DialogViewModel.DialogType) {
+        self.dialogType = dialogType
+        //Keep the system font for enjoyment dialog.
+        if self.dialogType == .enjoymentDialog {
+            self.titleFont = .preferredFont(forTextStyle: .headline)
+        }
+        self.collapseImageBottomPadding = collapseImageBottomPadding
         self.blurEffect = UIBlurEffect(style: .systemChromeMaterial)
-
         self.backgroundView = UIVisualEffectView(effect: self.blurEffect)
         self.headerImageView = UIImageView()
+        self.headerImageAlternateLabel = UILabel()
         self.textScrollView = UIScrollView()
         self.textContentView = UIView()
         self.titleLabel = UILabel()
         self.messageLabel = UILabel()
         self.gradientView = GradientView()
-        self.middleSeparator = Self.buildSeparatorView(with: self.blurEffect)
+        self.mainSeparator = Self.buildSeparatorView(with: self.blurEffect)
         self.buttonStackView = UIStackView()
 
         super.init(frame: frame)
@@ -106,14 +145,10 @@ public class DialogView: UIView {
         self.messageLabel.textColor = self.messageTextColor
         self.messageLabel.font = self.messageFont
 
-        if let headerImage = self.headerImage {
-            self.headerImageView.image = headerImage
-            self.headerImageView.isHidden = false
+        self.mainSeparator.backgroundColor = self.separatorColor
 
-            NSLayoutConstraint.deactivate([self.headerImageViewHeightConstraint])
-            NSLayoutConstraint.activate([
-                self.headerImageView.heightAnchor.constraint(equalTo: self.headerImageView.widthAnchor, multiplier: headerImage.size.height / headerImage.size.width)
-            ])
+        for separator in self.buttonSeparators {
+            separator.backgroundColor = self.separatorColor
         }
     }
 
@@ -131,12 +166,14 @@ public class DialogView: UIView {
 
         self.gradientView.isHidden = self.textScrollView.contentSize.height <= self.textScrollView.bounds.height
 
-        self.middleSeparator.isHidden = self.separatorsAreHidden
-        self.middleSeparator.backgroundColor = self.separatorColor
+        self.mainSeparator.isHidden = self.separatorsAreHidden
+        self.mainSeparator.backgroundColor = self.separatorColor
+        self.mainSeparator.contentView.backgroundColor = .white
 
         for separator in self.buttonSeparators {
             separator.isHidden = self.separatorsAreHidden
             separator.backgroundColor = self.separatorColor
+            separator.contentView.backgroundColor = .white
         }
 
         self.layer.cornerRadius = self.cornerRadius
@@ -156,7 +193,7 @@ public class DialogView: UIView {
     private func layoutSeparators() {
         let separatorThickness = 1.0 / self.traitCollection.displayScale
 
-        self.middleSeparator.frame = .init(x: 0, y: self.buttonStackView.frame.minY - separatorThickness - self.buttonInset.top, width: self.bounds.width, height: separatorThickness)
+        self.mainSeparator.frame = .init(x: 0, y: self.buttonStackView.frame.minY - separatorThickness - self.buttonInset.top, width: self.bounds.width, height: separatorThickness)
 
         for (index, separator) in self.buttonSeparators.enumerated() {
             switch self.buttonStackView.axis {
@@ -187,10 +224,9 @@ public class DialogView: UIView {
 
     private func configure() {
         self.addSubview(self.backgroundView)
-        self.backgroundView.contentView.addSubview(self.headerImageView)
         self.backgroundView.contentView.addSubview(self.textScrollView)
         self.backgroundView.contentView.addSubview(self.gradientView)
-        self.backgroundView.contentView.addSubview(self.middleSeparator)
+        self.backgroundView.contentView.addSubview(self.mainSeparator)
         self.backgroundView.contentView.addSubview(self.buttonStackView)
 
         // Cutting corners
@@ -200,6 +236,7 @@ public class DialogView: UIView {
         self.layer.masksToBounds = true
 
         self.headerImageView.contentMode = .scaleAspectFit
+        self.headerImageView.clipsToBounds = true
 
         self.buttonStackView.alignment = .center
         self.buttonStackView.distribution = .fillEqually
@@ -210,8 +247,10 @@ public class DialogView: UIView {
     private func configureScrollView() {
         self.textScrollView.addSubview(self.textContentView)
 
+        self.textContentView.addSubview(self.headerImageView)
         self.textContentView.addSubview(self.titleLabel)
         self.textContentView.addSubview(self.messageLabel)
+        self.textContentView.addSubview(self.headerImageAlternateLabel)
 
         self.titleLabel.numberOfLines = 0
         self.titleLabel.textAlignment = .center
@@ -223,17 +262,19 @@ public class DialogView: UIView {
         self.messageLabel.lineBreakMode = .byWordWrapping
         self.messageLabel.adjustsFontForContentSizeCategory = true
 
+        self.headerImageAlternateLabel.numberOfLines = 0
+        self.headerImageAlternateLabel.textAlignment = .center
+        self.headerImageAlternateLabel.lineBreakMode = .byWordWrapping
+        self.headerImageAlternateLabel.adjustsFontForContentSizeCategory = true
+
         self.setScrollViewConstraints()
     }
 
     private func setConstraints() {
         self.backgroundView.translatesAutoresizingMaskIntoConstraints = false
-        self.headerImageView.translatesAutoresizingMaskIntoConstraints = false
         self.textScrollView.translatesAutoresizingMaskIntoConstraints = false
         self.gradientView.translatesAutoresizingMaskIntoConstraints = false
         self.buttonStackView.translatesAutoresizingMaskIntoConstraints = false
-
-        self.headerImageViewHeightConstraint = self.headerImageView.heightAnchor.constraint(equalToConstant: 0)
 
         self.buttonStackViewTopConstraint = self.buttonStackView.topAnchor.constraint(equalTo: self.textScrollView.bottomAnchor)
         self.buttonStackViewLeftConstraint = self.buttonStackView.leftAnchor.constraint(equalTo: self.leftAnchor)
@@ -246,12 +287,7 @@ public class DialogView: UIView {
             self.trailingAnchor.constraint(equalTo: self.backgroundView.trailingAnchor),
             self.bottomAnchor.constraint(equalTo: self.backgroundView.bottomAnchor),
 
-            self.headerImageView.topAnchor.constraint(equalTo: self.backgroundView.topAnchor),
-            self.headerImageView.leadingAnchor.constraint(equalTo: self.backgroundView.leadingAnchor),
-            self.backgroundView.trailingAnchor.constraint(equalTo: self.headerImageView.trailingAnchor),
-            self.headerImageViewHeightConstraint,
-
-            self.textScrollView.topAnchor.constraint(equalTo: self.headerImageView.bottomAnchor),
+            self.textScrollView.topAnchor.constraint(equalTo: self.backgroundView.topAnchor),
             self.textScrollView.leadingAnchor.constraint(equalTo: self.backgroundView.leadingAnchor),
             self.backgroundView.trailingAnchor.constraint(equalTo: self.textScrollView.trailingAnchor),
 
@@ -268,8 +304,19 @@ public class DialogView: UIView {
     }
 
     private func setScrollViewConstraints() {
+
+        if self.collapseImageBottomPadding {
+            self.topSpacingIdealConstraintConstant = 0
+            self.topSpacingMinConstraintConstant = 0
+            self.topSpacingMaxConstraintConstant = 0
+            self.titleBottomConstraintConstant = 0
+            self.messageBottomConstraintConstant = 0
+        }
+
         self.textScrollView.translatesAutoresizingMaskIntoConstraints = false
         self.textContentView.translatesAutoresizingMaskIntoConstraints = false
+        self.headerImageView.translatesAutoresizingMaskIntoConstraints = false
+        self.headerImageAlternateLabel.translatesAutoresizingMaskIntoConstraints = false
         self.titleLabel.translatesAutoresizingMaskIntoConstraints = false
         self.messageLabel.translatesAutoresizingMaskIntoConstraints = false
 
@@ -277,17 +324,27 @@ public class DialogView: UIView {
         textSpacingIdealConstraint.priority = UILayoutPriority(746)
         let textSpacingMinConstraint = self.messageLabel.firstBaselineAnchor.constraint(greaterThanOrEqualTo: self.titleLabel.lastBaselineAnchor, constant: 17)
         let textSpacingMaxConstraint = self.messageLabel.firstBaselineAnchor.constraint(lessThanOrEqualTo: self.titleLabel.lastBaselineAnchor, constant: 20)
+        textSpacingMaxConstraint.priority = UILayoutPriority(999)
 
-        let topSpacingIdealConstraint = self.titleLabel.firstBaselineAnchor.constraint(equalTo: self.textContentView.topAnchor, constant: 36)
-        topSpacingIdealConstraint.priority = UILayoutPriority(748)
-        let topSpacingMinConstraint = self.titleLabel.firstBaselineAnchor.constraint(greaterThanOrEqualTo: self.textContentView.topAnchor, constant: 22.3333)
-        let topSpacingMaxConstraint = self.titleLabel.firstBaselineAnchor.constraint(lessThanOrEqualTo: self.textContentView.topAnchor, constant: 36)
+        self.topSpacingIdealConstraint = self.titleLabel.firstBaselineAnchor.constraint(greaterThanOrEqualTo: self.headerImageView.bottomAnchor, constant: self.topSpacingIdealConstraintConstant)
+        self.topSpacingIdealConstraint.priority = UILayoutPriority(748)
+
+        self.topSpacingMinConstraint = self.titleLabel.firstBaselineAnchor.constraint(greaterThanOrEqualTo: self.headerImageView.bottomAnchor, constant: self.topSpacingMinConstraintConstant)
+        self.topSpacingMinConstraint.priority = UILayoutPriority(999)
+        self.topSpacingMaxConstraint = self.titleLabel.firstBaselineAnchor.constraint(lessThanOrEqualTo: self.headerImageView.bottomAnchor, constant: self.topSpacingMaxConstraintConstant)
+        self.topSpacingMaxConstraint.priority = UILayoutPriority(749)
 
         let heightExpanderConstraint = self.textScrollView.heightAnchor.constraint(equalTo: self.textContentView.heightAnchor)
         heightExpanderConstraint.priority = UILayoutPriority(749)
 
-        self.titleBottomConstraint = self.textContentView.bottomAnchor.constraint(equalTo: self.titleLabel.lastBaselineAnchor, constant: 24)
-        self.messageBottomConstraint = self.textContentView.bottomAnchor.constraint(equalTo: self.messageLabel.lastBaselineAnchor, constant: 24)
+        self.titleBottomConstraint = self.textContentView.bottomAnchor.constraint(equalTo: self.titleLabel.lastBaselineAnchor, constant: self.titleBottomConstraintConstant)
+        self.messageBottomConstraint = self.textContentView.bottomAnchor.constraint(equalTo: self.messageLabel.lastBaselineAnchor, constant: self.messageBottomConstraintConstant)
+
+        self.headerImageViewHeightConstraint = self.headerImageView.heightAnchor.constraint(equalToConstant: 0)
+        self.headerImageViewTopConstraint = self.headerImageView.topAnchor.constraint(equalTo: self.textContentView.topAnchor)
+        self.headerImageViewLeadingConstraint = self.headerImageView.leadingAnchor.constraint(equalTo: self.textContentView.leadingAnchor)
+        self.headerImageViewTrailingConstraint = self.textContentView.trailingAnchor.constraint(equalTo: self.headerImageView.trailingAnchor)
+        self.headerImageViewBottomConstraint = self.titleLabel.topAnchor.constraint(equalTo: self.headerImageView.bottomAnchor)
 
         NSLayoutConstraint.activate([
             self.textContentView.topAnchor.constraint(equalTo: self.textScrollView.topAnchor),
@@ -297,12 +354,24 @@ public class DialogView: UIView {
             self.textContentView.rightAnchor.constraint(equalTo: self.textScrollView.rightAnchor),
             self.textContentView.leftAnchor.constraint(equalTo: self.textScrollView.leftAnchor),
 
-            topSpacingIdealConstraint, topSpacingMinConstraint, topSpacingMaxConstraint,
+            self.headerImageViewTopConstraint,
+            self.headerImageViewLeadingConstraint,
+            self.headerImageViewTrailingConstraint,
+
+            self.headerImageAlternateLabel.topAnchor.constraint(equalToSystemSpacingBelow: self.textContentView.topAnchor, multiplier: 1),
+            self.headerImageAlternateLabel.leadingAnchor.constraint(greaterThanOrEqualToSystemSpacingAfter: self.textContentView.leadingAnchor, multiplier: 1),
+            self.textContentView.trailingAnchor.constraint(greaterThanOrEqualToSystemSpacingAfter: self.headerImageAlternateLabel.trailingAnchor, multiplier: 1),
+            self.headerImageAlternateLabel.centerXAnchor.constraint(equalTo: self.textContentView.centerXAnchor),
+            self.headerImageAlternateLabel.heightAnchor.constraint(equalTo: self.headerImageView.heightAnchor, multiplier: 1),
+
+            self.topSpacingIdealConstraint, topSpacingMinConstraint, topSpacingMaxConstraint,
+            self.titleLabel.topAnchor.constraint(greaterThanOrEqualTo: self.headerImageView.bottomAnchor),
 
             self.titleLabel.centerXAnchor.constraint(equalTo: self.textContentView.centerXAnchor),
             self.titleLabel.widthAnchor.constraint(equalTo: self.textContentView.widthAnchor, constant: -32),
 
             textSpacingIdealConstraint, textSpacingMinConstraint, textSpacingMaxConstraint,
+            self.messageLabel.topAnchor.constraint(greaterThanOrEqualTo: self.titleLabel.bottomAnchor),
 
             self.messageLabel.centerXAnchor.constraint(equalTo: self.textContentView.centerXAnchor),
             self.messageLabel.widthAnchor.constraint(equalTo: self.textContentView.widthAnchor, constant: -32),
@@ -312,11 +381,7 @@ public class DialogView: UIView {
     }
 
     private static func buildSeparatorView(with blurEffect: UIBlurEffect) -> UIVisualEffectView {
-        let separatorView: UIVisualEffectView
-
-        separatorView = UIVisualEffectView(effect: UIVibrancyEffect(blurEffect: blurEffect, style: .separator))
-
-        separatorView.contentView.backgroundColor = .white
+        let separatorView = UIVisualEffectView(effect: UIVibrancyEffect(blurEffect: blurEffect, style: .separator))
 
         return separatorView
     }